feat(theme-docs): filter releases by releases setting
This commit add `releases` setting which can be used to filter the releases by: `preRelease`, `channel` and `tag`.

- `preRelease`: filter based on `prerelease` flag.
- `channel`: filter based on `target_commitish` value
- `tag`: filter based on `tag_name`.

The `preRelease` option is useful when you want to make different docs for stable and pre releases.

The `channel` option is really useful when CI make releases based on branch like `alpha`, `beta`, `next`. Making it easy to make docs for specific branches. Can be either a string, or an array of strings.

The `tag` option make it easy to filter by the tag name. For example, `v2.*.*-alpha` will return alpha releases of v2. Can be either a string, or an array of strings.