Fixes #1780
When a user uses the [path explicitly](https://content.nuxtjs.org/api/components/content-doc/#explicit-path) and the file does not exist, it should throw a 404. But instead, that `if (!content)` block doesn't execute at all since inside the `if(path)` block before that returns a valid response even if the `content` variable is undefined.