Update 2.content-renderer.md
Attempting to clarify something that's cause me some confusion. 

Prior to this change, the documentation implied that setting this to "true" causes the excerpt to appear and setting it to "false" would cause the excerpt NOT to appear. 

In actuality, setting this prop to "true" or "false" does not impact the excerpt at all. It only impacts the remaining document after the excerpt.

Naming things is hard.