Update ContentSlot Documentation
As far as I can tell custom ContentSlot components need to be in the `components/content` directory. I spent a whole day trying to figure out how to get these working, then simply moved the components to the content directory and it worked. I double checked this functionality with the stackblitz environment linked below. I tried to find the reference code that would actually enforce this, but I couldn't find it.

https://stackblitz.com/edit/github-syrfpd-jmspgd?file=README.md